Check for null/empty values before clustering
Having null or empty value is a case not considered in the
algorithm. This will lead to a infinite loop since some
features will not add to any region.
If any of the attributes for the clutering or threshold appear
as null/empty values, the plugin will display an error with the
first feature that failed this condition.